refactor(product): extract API layer into dojo/product/api/ [product Phase 6,7,8,9]
