+++ dojo/api_v2/views.py
@@ -56,7 +56,6 @@ from dojo.filters import (
     ApiDojoMetaFilter,
     ApiEndpointFilter,
     ApiFindingFilter,
-    ApiProductFilter,
     ApiRiskAcceptanceFilter,
     ApiTemplateFindingFilter,
     ApiUserFilter,
@@ -93,7 +92,6 @@ from dojo.models import (
     NoteHistory,
     Notes,
     Product,
-    Product_API_Scan_Configuration,
     Regulation,
     Risk_Acceptance,
     SLA_Configuration,
@@ -111,7 +109,6 @@ from dojo.product.queries import (
     get_authorized_app_analysis,
     get_authorized_dojo_meta,
     get_authorized_languages,
-    get_authorized_product_api_scan_configurations,
     get_authorized_products,
 )
 from dojo.query_utils import build_count_subquery
@@ -127,12 +124,10 @@ from dojo.tool_product.queries import get_authorized_tool_product_settings
 from dojo.user.authentication import reset_token_for_user
 from dojo.user.utils import get_configuration_permissions_codenames
 from dojo.utils import (
-    async_delete,
     generate_file_response,
     get_celery_queue_details,
     get_celery_queue_length,
     get_celery_worker_status,
-    get_setting,
     get_system_setting,
     process_tag_notifications,
     purge_celery_queue,
@@ -1258,31 +1253,6 @@ class SonarqubeIssueTransitionViewSet(
 
 # Authorization: object-based
 @extend_schema_view(**schema_with_prefetch())
-class ProductAPIScanConfigurationViewSet(
-    PrefetchDojoModelViewSet,
-):
-    serializer_class = serializers.ProductAPIScanConfigurationSerializer
-    queryset = Product_API_Scan_Configuration.objects.none()
-    filter_backends = (DjangoFilterBackend,)
-    filterset_fields = [
-        "id",
-        "product",
-        "tool_configuration",
-        "service_key_1",
-        "service_key_2",
-        "service_key_3",
-    ]
-    permission_classes = (
-        IsAuthenticated,
-        permissions.UserHasProductAPIScanConfigurationPermission,
-    )
-
-    def get_queryset(self):
-        return get_authorized_product_api_scan_configurations(
-            "view",
-        )
-
-
 # Authorization: object-based
 # @extend_schema_view(**schema_with_prefetch())
 # Nested models with prefetch make the response schema too long for Swagger UI
@@ -1374,86 +1344,6 @@ class DojoMetaViewSet(
             else:
                 msg = f"Metadata {metadata.get('name')} not found for object."
                 raise ValidationError(msg)
-
-
-@extend_schema_view(**schema_with_prefetch())
-class ProductViewSet(
-    prefetch.PrefetchListMixin,
-    prefetch.PrefetchRetrieveMixin,
-    mixins.CreateModelMixin,
-    mixins.DestroyModelMixin,
-    mixins.UpdateModelMixin,
-    viewsets.GenericViewSet,
-    dojo_mixins.DeletePreviewModelMixin,
-):
-    serializer_class = serializers.ProductSerializer
-    queryset = Product.objects.none()
-    filter_backends = (DjangoFilterBackend,)
-    filterset_class = ApiProductFilter
-    permission_classes = (
-        IsAuthenticated,
-        permissions.UserHasProductPermission,
-    )
-
-    def get_queryset(self):
-        return get_authorized_products("view").distinct()
-
-    def destroy(self, request, *args, **kwargs):
-        instance = self.get_object()
-        if get_setting("ASYNC_OBJECT_DELETE"):
-            async_del = async_delete()
-            async_del.delete(instance)
-        else:
-            with Endpoint.allow_endpoint_init():  # TODO: Delete this after the move to Locations
-                instance.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
-    # def list(self, request):
-    #     # Note the use of `get_queryset()` instead of `self.queryset`
-    #     queryset = self.get_queryset()
-    #     serializer = self.serializer_class(queryset, many=True)
-    #     return Response(serializer.data)
-
-    @extend_schema(
-        request=serializers.ReportGenerateOptionSerializer,
-        responses={status.HTTP_200_OK: serializers.ReportGenerateSerializer},
-    )
-    @action(
-        detail=True, methods=["post"],
-        # IsAuthenticated only: report generation requires View permission,
-        # enforced by the permission-filtered get_queryset(). The viewset's
-        # permission_classes would check Edit (POST), which is too restrictive.
-        permission_classes=[IsAuthenticated],
-    )
-    def generate_report(self, request, pk=None):
-        product = self.get_object()
-
-        options = {}
-        # prepare post data
-        report_options = serializers.ReportGenerateOptionSerializer(
-            data=request.data,
-        )
-        if report_options.is_valid():
-            options["include_finding_notes"] = report_options.validated_data[
-                "include_finding_notes"
-            ]
-            options["include_finding_images"] = report_options.validated_data[
-                "include_finding_images"
-            ]
-            options[
-                "include_executive_summary"
-            ] = report_options.validated_data["include_executive_summary"]
-            options[
-                "include_table_of_contents"
-            ] = report_options.validated_data["include_table_of_contents"]
-        else:
-            return Response(
-                report_options.errors, status=status.HTTP_400_BAD_REQUEST,
-            )
-
-        data = report_generate(request, product, options)
-        report = serializers.ReportGenerateSerializer(data)
-        return Response(report.data)
 
 
 # Authorization: authenticated, configuration
