+++ dojo/product/api/__init__.py
@@ -1,0 +1,1 @@
+path = "products"  # noqa: RUF067
