+++ unittests/test_rest_framework.py
@@ -55,8 +55,6 @@ from dojo.api_v2.views import (
     LanguageViewSet,
     NotesViewSet,
     NoteTypeViewSet,
-    ProductAPIScanConfigurationViewSet,
-    ProductViewSet,
     RiskAcceptanceViewSet,
     SonarqubeIssueViewSet,
     ToolConfigurationsViewSet,
@@ -113,6 +111,7 @@ from dojo.notifications.api.views import NotificationsViewSet, NotificationWebho
 from dojo.organization.api.views import (
     OrganizationViewSet,
 )
+from dojo.product.api.views import ProductAPIScanConfigurationViewSet, ProductViewSet
 from dojo.product_type.api.views import ProductTypeViewSet
 from dojo.test.api.views import TestsViewSet, TestTypesViewSet
 from dojo.url.api.views import URLViewSet
